sdks/python-sdk/requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.3
jsons==1.6.3
waitress==2.1.2
paho-mqtt==1.5.1